requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.4
python-dateutil==2.8.2
sqlalchemy==2.0.25
mangum==0.17.0
psycopg2-binary==2.9.9
